test(contracts): re-organize smart contract tests in separate files
Currently all tests related to a Poll were inside one file (MACI.test) - these tests have been moved
into their respective files, re organized and cleaned up, and a small number of test cases were
added which were missing.
