+++ contracts/package.json
@@ -14,6 +14,9 @@
     "docs": "hardhat docgen",
     "test": "hardhat test",
     "test-maci": "hardhat test ./tests/MACI.test.ts",
+    "test-poll": "hardhat test ./tests/Poll.test.ts",
+    "test-messageProcessor": "hardhat test ./tests/MessageProcessor.test.ts",
+    "test-tally": "hardhat test ./tests/Tally.test.ts",
     "test-hasher": "hardhat test ./tests/Hasher.test.ts",
     "test-utilities": "hardhat test ./tests/Utilities.test.ts",
     "test-signupGatekeeper": "hardhat test ./tests/SignUpGatekeeper.test.ts",
