+++ contracts/tests/MACI.test.ts
@@ -1,115 +1,43 @@
 /* eslint-disable no-underscore-dangle */
 import { expect } from "chai";
 import { AbiCoder, BaseContract, BigNumberish, Signer } from "ethers";
-import {
-  MaciState,
-  genProcessVkSig,
-  MaxValues,
-  TreeDepths,
-  packProcessMessageSmallVals,
-  packTallyVotesSmallVals,
-  Poll,
-} from "maci-core";
-import { G1Point, G2Point, NOTHING_UP_MY_SLEEVE } from "maci-crypto";
-import { PCommand, VerifyingKey, Keypair, PubKey, Message } from "maci-domainobjs";
+import { EthereumProvider } from "hardhat/types";
+import { MaciState, genProcessVkSig, STATE_TREE_DEPTH } from "maci-core";
+import { NOTHING_UP_MY_SLEEVE } from "maci-crypto";
+import { VerifyingKey, Keypair, PubKey, Message } from "maci-domainobjs";
 
 import type { IVerifyingKeyStruct } from "../ts/types";
-import type { EthereumProvider } from "hardhat/types";
 
 import { parseArtifact, getDefaultSigner } from "../ts/deploy";
 import { deployTestContracts } from "../ts/utils";
+import { AccQueueQuinaryMaci, MACI, VkRegistry, Poll as PollContract } from "../typechain-types";
+
 import {
-  AccQueueQuinaryMaci,
-  AccQueue,
-  MACI,
-  MessageProcessor,
-  Tally,
-  VkRegistry,
-  Poll as PollContract,
-} from "../typechain-types";
-
-import { timeTravel } from "./utils";
-
-const STATE_TREE_DEPTH = 10;
-const STATE_TREE_ARITY = 5;
-const MESSAGE_TREE_DEPTH = 4;
-const MESSAGE_TREE_SUBDEPTH = 2;
-
-const coordinator = new Keypair();
-const [pollAbi] = parseArtifact("Poll");
-const [accQueueQuinaryMaciAbi] = parseArtifact("AccQueueQuinaryMaci");
-
-const testProcessVk = new VerifyingKey(
-  new G1Point(BigInt(0), BigInt(1)),
-  new G2Point([BigInt(2), BigInt(3)], [BigInt(4), BigInt(5)]),
-  new G2Point([BigInt(6), BigInt(7)], [BigInt(8), BigInt(9)]),
-  new G2Point([BigInt(10), BigInt(11)], [BigInt(12), BigInt(13)]),
-  [new G1Point(BigInt(14), BigInt(15)), new G1Point(BigInt(16), BigInt(17))],
-);
-
-const testTallyVk = new VerifyingKey(
-  new G1Point(BigInt(0), BigInt(1)),
-  new G2Point([BigInt(2), BigInt(3)], [BigInt(4), BigInt(5)]),
-  new G2Point([BigInt(6), BigInt(7)], [BigInt(8), BigInt(9)]),
-  new G2Point([BigInt(10), BigInt(11)], [BigInt(12), BigInt(13)]),
-  [new G1Point(BigInt(14), BigInt(15)), new G1Point(BigInt(16), BigInt(17))],
-);
-
-const compareVks = (vk: VerifyingKey, vkOnChain: VerifyingKey) => {
-  expect(vk.ic.length).to.eq(vkOnChain.ic.length);
-  for (let i = 0; i < vk.ic.length; i += 1) {
-    expect(vk.ic[i].x.toString()).to.eq(vkOnChain.ic[i].x.toString());
-    expect(vk.ic[i].y.toString()).to.eq(vkOnChain.ic[i].y.toString());
-  }
-  expect(vk.alpha1.x.toString()).to.eq(vkOnChain.alpha1.x.toString());
-  expect(vk.alpha1.y.toString()).to.eq(vkOnChain.alpha1.y.toString());
-  expect(vk.beta2.x[0].toString()).to.eq(vkOnChain.beta2.x[0].toString());
-  expect(vk.beta2.x[1].toString()).to.eq(vkOnChain.beta2.x[1].toString());
-  expect(vk.beta2.y[0].toString()).to.eq(vkOnChain.beta2.y[0].toString());
-  expect(vk.beta2.y[1].toString()).to.eq(vkOnChain.beta2.y[1].toString());
-  expect(vk.delta2.x[0].toString()).to.eq(vkOnChain.delta2.x[0].toString());
-  expect(vk.delta2.x[1].toString()).to.eq(vkOnChain.delta2.x[1].toString());
-  expect(vk.delta2.y[0].toString()).to.eq(vkOnChain.delta2.y[0].toString());
-  expect(vk.delta2.y[1].toString()).to.eq(vkOnChain.delta2.y[1].toString());
-  expect(vk.gamma2.x[0].toString()).to.eq(vkOnChain.gamma2.x[0].toString());
-  expect(vk.gamma2.x[1].toString()).to.eq(vkOnChain.gamma2.x[1].toString());
-  expect(vk.gamma2.y[0].toString()).to.eq(vkOnChain.gamma2.y[0].toString());
-  expect(vk.gamma2.y[1].toString()).to.eq(vkOnChain.gamma2.y[1].toString());
-};
-
-const users = [new Keypair(), new Keypair(), new Keypair()];
-
-const signUpTxOpts = { gasLimit: 300000 };
-
-const maciState = new MaciState(STATE_TREE_DEPTH);
-
-// Poll parameters
-const duration = 15;
-const maxValues: MaxValues = {
-  maxMessages: 25,
-  maxVoteOptions: 25,
-};
-
-const treeDepths: TreeDepths = {
-  intStateTreeDepth: 1,
-  messageTreeDepth: MESSAGE_TREE_DEPTH,
-  messageTreeSubDepth: MESSAGE_TREE_SUBDEPTH,
-  voteOptionTreeDepth: 2,
-};
-
-const messageBatchSize = 25;
-const tallyBatchSize = STATE_TREE_ARITY ** treeDepths.intStateTreeDepth;
-
-const initialVoiceCreditBalance = 100;
-let signer: Signer;
+  duration,
+  initialVoiceCreditBalance,
+  maxValues,
+  messageBatchSize,
+  tallyBatchSize,
+  testProcessVk,
+  testTallyVk,
+  treeDepths,
+} from "./constants";
+import { compareVks, timeTravel } from "./utils";
 
 describe("MACI", () => {
   let maciContract: MACI;
   let stateAqContract: AccQueueQuinaryMaci;
   let vkRegistryContract: VkRegistry;
-  let mpContract: MessageProcessor;
-  let tallyContract: Tally;
   let pollId: number;
+
+  const coordinator = new Keypair();
+  const [pollAbi] = parseArtifact("Poll");
+  const users = [new Keypair(), new Keypair(), new Keypair()];
+
+  let signer: Signer;
+
+  const maciState = new MaciState(STATE_TREE_DEPTH);
+  const signUpTxOpts = { gasLimit: 300000 };
 
   describe("Deployment", () => {
     before(async () => {
@@ -119,8 +47,6 @@ describe("MACI", () => {
       maciContract = r.maciContract;
       stateAqContract = r.stateAqContract;
       vkRegistryContract = r.vkRegistryContract;
-      mpContract = r.mpContract;
-      tallyContract = r.tallyContract;
     });
 
     it("MACI.stateTreeDepth should be correct", async () => {
@@ -178,19 +104,29 @@ describe("MACI", () => {
         ),
       ).to.be.revertedWithCustomError(maciContract, "MaciPubKeyLargerThanSnarkFieldSize");
     });
-  });
 
-  describe("Merging sign-ups should fail because of onlyPoll", () => {
-    it("coordinator should not be able to merge the signUp AccQueue", async () => {
-      await expect(maciContract.mergeStateAqSubRoots(0, 0, { gasLimit: 3000000 })).to.be.revertedWithCustomError(
-        maciContract,
-        "CallerMustBePoll",
-      );
+    it("should not allow to sign up more than the supported amount of users (5 ** stateTreeDepth)", async () => {
+      const stateTreeDepthTest = 1;
+      const maxUsers = 5 ** stateTreeDepthTest;
+      const maci = (await deployTestContracts(initialVoiceCreditBalance, stateTreeDepthTest, true)).maciContract;
+      const keypair = new Keypair();
+      for (let i = 0; i < maxUsers; i += 1) {
+        // eslint-disable-next-line no-await-in-loop
+        await maci.signUp(
+          keypair.pubKey.asContractParam(),
+          AbiCoder.defaultAbiCoder().encode(["uint256"], [1]),
+          AbiCoder.defaultAbiCoder().encode(["uint256"], [0]),
+        );
+      }
 
-      await expect(maciContract.mergeStateAq(0, { gasLimit: 3000000 })).to.be.revertedWithCustomError(
-        maciContract,
-        "CallerMustBePoll",
-      );
+      // the next signup should fail
+      await expect(
+        maci.signUp(
+          keypair.pubKey.asContractParam(),
+          AbiCoder.defaultAbiCoder().encode(["uint256"], [1]),
+          AbiCoder.defaultAbiCoder().encode(["uint256"], [0]),
+        ),
+      ).to.be.revertedWithCustomError(maci, "TooManySignups");
     });
   });
 
@@ -347,9 +283,19 @@ describe("MACI", () => {
       expect(Number(onChainBatchSizes.messageBatchSize)).to.eq(messageBatchSize);
       expect(Number(onChainBatchSizes.tallyBatchSize)).to.eq(tallyBatchSize);
     });
+
+    it("should prevent deploying a second poll before the first has finished", async () => {
+      await expect(
+        maciContract.deployPoll(duration, maxValues, treeDepths, coordinator.pubKey.asContractParam(), {
+          gasLimit: 8000000,
+        }),
+      )
+        .to.be.revertedWithCustomError(maciContract, "PreviousPollNotCompleted")
+        .withArgs(1);
+    });
   });
 
-  describe("Publish messages (vote + key-change)", () => {
+  describe("Merge sign-ups", () => {
     let pollContract: PollContract;
 
     before(async () => {
@@ -357,130 +303,20 @@ describe("MACI", () => {
       pollContract = new BaseContract(pollContractAddress, pollAbi, signer) as PollContract;
     });
 
-    it("should publish a message to the Poll contract", async () => {
-      const keypair = new Keypair();
-
-      const command = new PCommand(
-        BigInt(1),
-        keypair.pubKey,
-        BigInt(0),
-        BigInt(9),
-        BigInt(1),
-        BigInt(pollId),
-        BigInt(0),
+    it("coordinator should not be able to merge the signUp AccQueue", async () => {
+      await expect(maciContract.mergeStateAqSubRoots(0, 0, { gasLimit: 3000000 })).to.be.revertedWithCustomError(
+        maciContract,
+        "CallerMustBePoll",
       );
 
-      const signature = command.sign(keypair.privKey);
-      const sharedKey = Keypair.genEcdhSharedKey(keypair.privKey, coordinator.pubKey);
-      const message = command.encrypt(signature, sharedKey);
-      const tx = await pollContract.publishMessage(
-        message.asContractParam(),
-        keypair.pubKey.asContractParam() as { x: BigNumberish; y: BigNumberish },
+      await expect(maciContract.mergeStateAq(0, { gasLimit: 3000000 })).to.be.revertedWithCustomError(
+        maciContract,
+        "CallerMustBePoll",
       );
-      const receipt = await tx.wait();
-      expect(receipt?.status).to.eq(1);
-
-      maciState.polls[pollId].publishMessage(message, keypair.pubKey);
-    });
-
-    it("shold not publish a message after the voting period", async () => {
-      const dd = await pollContract.getDeployTimeAndDuration();
-      await timeTravel(signer.provider as unknown as EthereumProvider, Number(dd[0]) + 1);
-
-      const keypair = new Keypair();
-      const command = new PCommand(
-        BigInt(0),
-        keypair.pubKey,
-        BigInt(0),
-        BigInt(0),
-        BigInt(0),
-        BigInt(pollId),
-        BigInt(0),
-      );
-
-      const signature = command.sign(keypair.privKey);
-      const sharedKey = Keypair.genEcdhSharedKey(keypair.privKey, coordinator.pubKey);
-      const message = command.encrypt(signature, sharedKey);
-
-      await expect(
-        pollContract.publishMessage(
-          message.asContractParam(),
-          keypair.pubKey.asContractParam() as { x: BigNumberish; y: BigNumberish },
-          { gasLimit: 300000 },
-        ),
-      ).to.be.revertedWithCustomError(pollContract, "VotingPeriodOver");
-    });
-  });
-
-  describe("Merge messages", () => {
-    let pollContract: PollContract;
-    let messageAqContract: AccQueue;
-
-    beforeEach(async () => {
-      const pollContractAddress = await maciContract.getPoll(pollId);
-      pollContract = new BaseContract(pollContractAddress, pollAbi, signer) as PollContract;
-
-      const extContracts = await pollContract.extContracts();
-
-      const messageAqAddress = extContracts.messageAq;
-      messageAqContract = new BaseContract(messageAqAddress, accQueueQuinaryMaciAbi, signer) as AccQueue;
-    });
-
-    it("should revert if subtrees are not merged for StateAq", async () => {
-      await expect(pollContract.mergeMaciStateAq(0, { gasLimit: 4000000 })).to.be.revertedWithCustomError(
-        pollContract,
-        "StateAqSubtreesNeedMerge",
-      );
-    });
-
-    it("coordinator should be able to merge the message AccQueue", async () => {
-      let tx = await pollContract.mergeMessageAqSubRoots(0, {
-        gasLimit: 3000000,
-      });
-      let receipt = await tx.wait();
-      expect(receipt?.status).to.eq(1);
-
-      tx = await pollContract.mergeMessageAq({ gasLimit: 4000000 });
-      receipt = await tx.wait();
-      expect(receipt?.status).to.eq(1);
-    });
-
-    it("the message root must be correct", async () => {
-      const onChainMessageRoot = await messageAqContract.getMainRoot(MESSAGE_TREE_DEPTH);
-      const offChainMessageRoot = maciState.polls[pollId].messageTree.root;
-
-      expect(onChainMessageRoot.toString()).to.eq(offChainMessageRoot.toString());
-    });
-  });
-
-  describe("Tally votes (negative test)", () => {
-    it("tallyVotes() should fail as the messages have not been processed yet", async () => {
-      const pollContractAddress = await maciContract.getPoll(pollId);
-      await expect(
-        tallyContract.tallyVotes(pollContractAddress, await mpContract.getAddress(), 0, [0, 0, 0, 0, 0, 0, 0, 0]),
-      ).to.be.revertedWithCustomError(tallyContract, "ProcessingNotComplete");
-    });
-  });
-
-  describe("Process messages (negative test)", () => {
-    it("processMessages() should fail if the state AQ has not been merged", async () => {
-      const pollContractAddress = await maciContract.getPoll(pollId);
-
-      await expect(
-        mpContract.processMessages(pollContractAddress, 0, [0, 0, 0, 0, 0, 0, 0, 0]),
-      ).to.be.revertedWithCustomError(mpContract, "StateAqNotMerged");
-    });
-  });
-
-  describe("Merge sign-ups as the Poll", () => {
-    let pollContract: PollContract;
-
-    before(async () => {
-      const pollContractAddress = await maciContract.getPoll(pollId);
-      pollContract = new BaseContract(pollContractAddress, pollAbi, signer) as PollContract;
     });
 
     it("The Poll should be able to merge the signUp AccQueue", async () => {
+      await timeTravel(signer.provider as unknown as EthereumProvider, Number(duration) + 1);
       let tx = await pollContract.mergeMaciStateAqSubRoots(0, pollId, {
         gasLimit: 3000000,
       });
@@ -497,90 +333,6 @@ describe("MACI", () => {
     it("the state root must be correct", async () => {
       const onChainStateRoot = await stateAqContract.getMainRoot(STATE_TREE_DEPTH);
       expect(onChainStateRoot.toString()).to.eq(maciState.stateTree.root.toString());
-    });
-  });
-
-  describe("Process messages", () => {
-    let pollContract: PollContract;
-    let poll: Poll;
-    let generatedInputs: { newSbCommitment: bigint };
-
-    before(async () => {
-      const pollContractAddress = await maciContract.getPoll(pollId);
-      pollContract = new BaseContract(pollContractAddress, pollAbi, signer) as PollContract;
-
-      poll = maciState.polls[pollId];
-      generatedInputs = poll.processMessages(pollId) as typeof generatedInputs;
-    });
-
-    it("genProcessMessagesPackedVals() should generate the correct value", async () => {
-      const packedVals = packProcessMessageSmallVals(
-        BigInt(maxValues.maxVoteOptions),
-        BigInt(users.length),
-        0,
-        poll.messages.length,
-      );
-      const onChainPackedVals = BigInt(
-        await mpContract.genProcessMessagesPackedVals(await pollContract.getAddress(), 0, users.length),
-      );
-      expect(packedVals.toString(16)).to.eq(onChainPackedVals.toString(16));
-    });
-
-    it("processMessages() should update the state and ballot root commitment", async () => {
-      const pollContractAddress = await maciContract.getPoll(pollId);
-
-      // Submit the proof
-      const tx = await mpContract.processMessages(
-        pollContractAddress,
-        generatedInputs.newSbCommitment,
-        [0, 0, 0, 0, 0, 0, 0, 0],
-      );
-
-      const receipt = await tx.wait();
-      expect(receipt?.status).to.eq(1);
-
-      const processingComplete = await mpContract.processingComplete();
-      expect(processingComplete).to.eq(true);
-
-      const onChainNewSbCommitment = await mpContract.sbCommitment();
-      expect(generatedInputs.newSbCommitment).to.eq(onChainNewSbCommitment.toString());
-    });
-  });
-
-  describe("Tally votes", () => {
-    it("genTallyVotesPackedVals() should generate the correct value", async () => {
-      const onChainPackedVals = BigInt(await tallyContract.genTallyVotesPackedVals(users.length, 0, tallyBatchSize));
-      const packedVals = packTallyVotesSmallVals(0, tallyBatchSize, users.length);
-      expect(onChainPackedVals.toString()).to.eq(packedVals.toString());
-    });
-
-    it("tallyVotes() should update the tally commitment", async () => {
-      const poll = maciState.polls[pollId];
-      const generatedInputs = poll.tallyVotes() as { newTallyCommitment: bigint };
-
-      const pollContractAddress = await maciContract.getPoll(pollId);
-      const tx = await tallyContract.tallyVotes(
-        pollContractAddress,
-        await mpContract.getAddress(),
-        generatedInputs.newTallyCommitment,
-        [0, 0, 0, 0, 0, 0, 0, 0],
-      );
-
-      const receipt = await tx.wait();
-      expect(receipt?.status).to.eq(1);
-
-      const onChainNewTallyCommitment = await tallyContract.tallyCommitment();
-
-      expect(generatedInputs.newTallyCommitment).to.eq(onChainNewTallyCommitment.toString());
-
-      await expect(
-        tallyContract.tallyVotes(
-          pollContractAddress,
-          await mpContract.getAddress(),
-          generatedInputs.newTallyCommitment,
-          [0, 0, 0, 0, 0, 0, 0, 0],
-        ),
-      ).to.be.revertedWithCustomError(tallyContract, "AllBallotsTallied");
     });
   });
 });
